ekf2: move baro control logic to file
